+++ desert-core/src/main/scala/io/github/vigoo/desert/custom/BinaryDeserializerOps.scala
@@ -1,86 +1,71 @@
 package io.github.vigoo.desert.custom
 
 import io.github.vigoo.desert.TypeRegistry.RegisteredTypeId
-import io.github.vigoo.desert.internal.SerializerState
 import io.github.vigoo.desert.internal.SerializerState.{RefId, StringId}
-import io.github.vigoo.desert._
-import zio.prelude.fx.ZPure
-
-import scala.util.Try
+import io.github.vigoo.desert.{BinaryCodec, BinaryDeserializer, DesertException, DesertFailure}
 
 trait BinaryDeserializerOps {
-  final def getInput: Deser[BinaryInput]                              = ZPure.serviceWith(_.input)
-  final def getInputTypeRegistry: Deser[TypeRegistry]                 = ZPure.serviceWith(_.typeRegistry)
-  final def getDeserializerState: Deser[SerializerState]              = ZPure.get
-  final def setDeserializerState(state: SerializerState): Deser[Unit] = ZPure.set(state)
 
-  final def readByte(): Deser[Byte]                              = getInput.flatMap(input => Deser.fromEither(input.readByte()))
-  final def readShort(): Deser[Short]                            = getInput.flatMap(input => Deser.fromEither(input.readShort()))
-  final def readInt(): Deser[Int]                                = getInput.flatMap(input => Deser.fromEither(input.readInt()))
-  final def readVarInt(optimizeForPositive: Boolean): Deser[Int] =
-    getInput.flatMap(input => Deser.fromEither(input.readVarInt(optimizeForPositive)))
-  final def readLong(): Deser[Long]                              = getInput.flatMap(input => Deser.fromEither(input.readLong()))
-  final def readFloat(): Deser[Float]                            = getInput.flatMap(input => Deser.fromEither(input.readFloat()))
-  final def readDouble(): Deser[Double]                          = getInput.flatMap(input => Deser.fromEither(input.readDouble()))
-  final def readBytes(count: Int): Deser[Array[Byte]]            =
-    getInput.flatMap(input => Deser.fromEither(input.readBytes(count)))
-  final def readCompressedByteArray(): Deser[Array[Byte]]        =
-    getInput.flatMap(input => Deser.fromEither(input.readCompressedByteArray()))
-  final def read[T: BinaryDeserializer](): Deser[T]              = implicitly[BinaryDeserializer[T]].deserialize()
+  final def failDeserializerWith[T](failure: DesertFailure): Nothing = throw DesertException(failure)
 
-  final def readUnknown(): Deser[Any] =
-    for {
-      typeRegistry <- getInputTypeRegistry
-      typeId       <- readVarInt(optimizeForPositive = true).map(RegisteredTypeId.apply)
-      result       <- typeRegistry.forId(typeId) match {
-                        case Some(registration) =>
-                          registration.codec.deserialize()
-                        case None               =>
-                          failDeserializerWith(DesertFailure.InvalidTypeId(typeId))
-                      }
-    } yield result
+  final def readByte()(implicit ctx: DeserializationContext): Byte = ctx.env.input.readByte()
 
-  final def finishDeserializerWith[T](value: T): Deser[T]                    = Deser.fromEither(Right(value))
-  final def failDeserializerWith[T](failure: DesertFailure): Deser[T]        = Deser.fromEither(Left(failure))
-  final def deserializerFromTry[T](f: Try[T], failMessage: String): Deser[T] =
-    Deser.fromEither(f.toEither.left.map(failure => DesertFailure.DeserializationFailure(failMessage, Some(failure))))
+  final def readShort()(implicit ctx: DeserializationContext): Short = ctx.env.input.readShort()
 
-  final def getString(value: StringId): Deser[Option[String]] =
-    for {
-      state <- getDeserializerState
-    } yield state.stringsById.get(value)
+  final def readInt()(implicit ctx: DeserializationContext): Int = ctx.env.input.readInt()
 
-  final def storeReadString(value: String): Deser[Unit] =
-    for {
-      state        <- getDeserializerState
-      (newState, _) = state.storeString(value)
-      _            <- setDeserializerState(newState)
-    } yield ()
+  final def readVarInt(optimizeForPositive: Boolean)(implicit ctx: DeserializationContext): Int =
+    ctx.env.input.readVarInt(optimizeForPositive)
 
-  final def getRef(value: RefId): Deser[Option[AnyRef]] =
-    for {
-      state <- getDeserializerState
-    } yield state.refsById.get(value)
+  final def readLong()(implicit ctx: DeserializationContext): Long = ctx.env.input.readLong()
 
-  final def storeReadRef(value: AnyRef): Deser[Unit] =
-    for {
-      state        <- getDeserializerState
-      (newState, _) = state.storeRef(value)
-      _            <- setDeserializerState(newState)
-    } yield ()
+  final def readFloat()(implicit ctx: DeserializationContext): Float = ctx.env.input.readFloat()
 
-  // TODO: need the codec to be lazy?
-  def readRefOrValue[T <: AnyRef](storeReadReference: Boolean = true)(implicit codec: BinaryCodec[T]): Deser[T] =
-    readVarInt(optimizeForPositive = true).flatMap {
+  final def readDouble()(implicit ctx: DeserializationContext): Double = ctx.env.input.readDouble()
+
+  final def readBytes(count: Int)(implicit ctx: DeserializationContext): Array[Byte] =
+    ctx.env.input.readBytes(count)
+
+  final def readCompressedByteArray()(implicit ctx: DeserializationContext): Array[Byte] =
+    ctx.env.input.readCompressedByteArray()
+
+  final def read[T: BinaryDeserializer]()(implicit ctx: DeserializationContext): T =
+    implicitly[BinaryDeserializer[T]].deserialize()
+
+  final def readUnknown()(implicit ctx: DeserializationContext): Any = {
+    val typeId = RegisteredTypeId(readVarInt(optimizeForPositive = true))
+    ctx.env.typeRegistry.forId(typeId) match {
+      case Some(registration) =>
+        registration.codec.deserialize()
+      case None               =>
+        failDeserializerWith(DesertFailure.InvalidTypeId(typeId))
+    }
+  }
+
+  final def getString(value: StringId)(implicit ctx: DeserializationContext): Option[String] =
+    ctx.state.getStringById(value)
+
+  final def storeReadString(value: String)(implicit ctx: DeserializationContext): Unit =
+    ctx.state.storeString(value)
+
+  final def getRef(value: RefId)(implicit ctx: DeserializationContext): Option[AnyRef] =
+    ctx.state.getRefById(value)
+
+  final def storeReadRef(value: AnyRef)(implicit ctx: DeserializationContext): Unit =
+    ctx.state.storeRef(value)
+
+  def readRefOrValue[T <: AnyRef](
+      storeReadReference: Boolean = true
+  )(implicit codec: BinaryCodec[T], ctx: DeserializationContext): T =
+    readVarInt(optimizeForPositive = true) match {
       case 0  =>
-        for {
-          value <- read[T]()
-          _     <- if (storeReadReference) storeReadRef(value) else finishDeserializerWith(())
-        } yield value
+        val value = read[T]()
+        if (storeReadReference) storeReadRef(value)
+        value
       case id =>
-        getRef(RefId(id)).flatMap {
+        getRef(RefId(id)) match {
           case None        => failDeserializerWith(DesertFailure.InvalidRefId(RefId(id)))
-          case Some(value) => finishDeserializerWith(value.asInstanceOf[T])
+          case Some(value) => value.asInstanceOf[T]
         }
     }
 }