+++ desert-zio/src/main/scala/io/github/vigoo/desert/ziosupport/custom/package.scala
@@ -1,15 +1,13 @@
 package io.github.vigoo.desert.ziosupport
 
 import io.github.vigoo.desert._
-import io.github.vigoo.desert.custom.{getInput, getOutput}
+import io.github.vigoo.desert.custom.{DeserializationContext, SerializationContext, readBytes, writeBytes}
 import zio.Chunk
 
 package object custom {
-  final def writeByteChunk[A](chunk: Chunk[Byte]): Ser[Unit] = getOutput.flatMap { output =>
-    Ser.fromEither(output.writeBytes(chunk.toArray))
-  }
+  final def writeByteChunk[A](chunk: Chunk[Byte])(implicit ctx: SerializationContext): Unit =
+    writeBytes(chunk.toArray)
 
-  final def readByteChunk[A](count: Int): Deser[Chunk[Byte]] = getInput.flatMap { input =>
-    Deser.fromEither(input.readBytes(count).map(Chunk.fromArray[Byte]))
-  }
+  final def readByteChunk[A](count: Int)(implicit ctx: DeserializationContext): Chunk[Byte] =
+    Chunk.fromArray(readBytes(count))
 }